pyproject.toml
[project]
name = "nextclimb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0",
    "django-cors-headers>=4.9.0",
    "django-ninja>=1.5.1",
    "geopy>=2.4.1",
    "python-dotenv>=1.2.1",
    "stravalib>=2.4",
]
